requirements.txt
aioredis
discord.py
jishaku
pillow
